LayoutTests/fast/dom/window-const-variable-shadowing-expected.txt

CONSOLE MESSAGE: SyntaxError: Can't create duplicate variable that shadows a global property: 'location'
CONSOLE MESSAGE: SyntaxError: Can't create duplicate variable that shadows a global property: 'foo'
Tests various cases of const variable shadowing on Window.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Should throw because Window has a non-configurable location own property.
PASS sentinel is "good"

Should throw because window has an own property foo that is not configurable, even though its value is undefined.
PASS sentinel is "good"

Should work because Window's own bar property is configurable.
PASS bar is 3
PASS sentinel is "good"

Should work because dispatchEvent is not an own property, it comes from the prototype chain.
PASS dispatchEvent is 3
PASS sentinel is "good"
PASS successfullyParsed is true

TEST COMPLETE
